refactor(test): decomposing
